Update shop.py
I updated on_buy function. Before modifying, it recieve room data from
client to use at destination of emit. But if user send wrong room value
that is mismatched with item, currect room of item can not receive
information that item is selled. So I use user of item to destination of
emit, because it is correct room number.